backend/requirements.txt
Django>=5.2,<6.0
psycopg2-binary>=2.9
# For REST API
djangorestframework>=3.15
# For JWT Auth
djangorestframework-simplejwt>=5.3
# For S3 storage
boto3>=1.34
django-storages>=1.14
# For CORS
django-cors-headers>=4.3
# For environment variables
python-dotenv>=1.0
# For file uploads and image handling (optional)
Pillow>=10.0 